[UPDATE] Fix rename to flags
